[FIx] - #126 TODO 채팅 입력 필드 클릭시 채팅 메세지 리스트도 함께 위로 올리기
